regions/system/library/python/config/pyproject.toml
[project]
name = "k1s0-config"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pydantic>=2.8", "pyyaml>=6.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/k1s0_config"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
fail_under = 90
